Test01

+++ scripts/package.sh
@@ -4,26 +4,32 @@ mkdir -p /workspace/package/DLPTemplates/tests/utils
 mkdir -p /workspace/package/DLPTemplates/tests/it-test
 cd /workspace/DLPTemplates/
 cp -R scripts/ /workspace/package/DLPTemplates/scripts/
-cp tests/* /workspace/package/DLPTemplates/tests/
-cp cloudbuild* /workspace/package/DLPTemplates/
+cp -R tests/* /workspace/package/DLPTemplates/tests/
 input="/workspace/dlp-diff.txt"
 cd /workspace/DLPTemplates
 while IFS= read -r line
 do
   subfolder=$(echo "$line" | cut -d '/' -f 1)
   if [[ "$subfolder" == "deIdentify_templates" ]]; then
-    cp "$line" "/workspace/package/DLPTemplates/deIdentify_templates/"
-    template_file=`echo "${line}" | awk -F/ '{print $NF}'`
-    it_test_file="${template_file%.*}"_test.py
-    cp "tests/it-test/$it_test_file" "/workspace/package/DLPTemplates/tests/it-test/"
+    if [[ -f "$line" ]]; then
+      cp "$line" "/workspace/package/DLPTemplates/deIdentify_templates/"
+      template_file=`echo "${line}" | awk -F/ '{print $NF}'`
+      it_test_file="${template_file%.*}"_test.py
+      cp "tests/it-test/$it_test_file" "/workspace/package/DLPTemplates/tests/it-test/"
+    else
+      echo "$line" | cut -d '/' -f 2 | sed -e 's/.json//g' >> "/workspace/package/DLPTemplates/deIdentify_templates_to_delete"
+    fi
   fi
   if [[ "$subfolder" == "inspect_templates" ]]; then
-    cp "$line" "/workspace/package/DLPTemplates/inspect_templates/"
-    template_file=`echo "${line}" | awk -F/ '{print $NF}'`
-    it_test_file="${template_file%.*}"_test.py
-    cp "tests/it-test/$it_test_file" "/workspace/package/DLPTemplates/tests/it-test/"
+    if [[ -f "$line" ]]; then
+      cp "$line" "/workspace/package/DLPTemplates/inspect_templates/"
+      template_file=`echo "${line}" | awk -F/ '{print $NF}'`
+      it_test_file="${template_file%.*}"_test.py
+      cp "tests/it-test/$it_test_file" "/workspace/package/DLPTemplates/tests/it-test/"
+    else
+      echo "$line" | cut -d '/' -f 2 | sed -e 's/.json//g' >> "/workspace/package/DLPTemplates/inspect_templates_to_delete"
+    fi
   fi
 done < "$input"
 cd /workspace/package/
-ls
 zip -r DLPTemplates.zip DLPTemplates/
